# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: string;
  name: string;
  phone: string;
  email?: string;
  dob?: string;
  notes?: string;
  totalSpend?: number;
  totalVisits?: number;
  createdAt?: string;
}

export interface CustomerListResponse {
  customers: Customer[];
}

export interface CreateCustomerBody {
  name: string;
  phone: string;
  email?: string;
  dob?: string;
  notes?: string;
}

export interface Appointment {
  id: string;
  customerId?: string;
  customerName?: string;
  customerPhone?: string;
  staffId: string;
  staffName?: string;
  serviceId: string;
  serviceName?: string;
  serviceCategory?: string;
  duration?: number;
  appointmentDate: string;
  appointmentTime: string;
  status: string;
  notes?: string;
  createdAt?: string;
}

export interface CustomerDetail {
  id: string;
  name: string;
  phone: string;
  email?: string;
  dob?: string;
  notes?: string;
  totalSpend?: number;
  totalVisits?: number;
  appointments?: Appointment[];
}

export interface Staff {
  id: string;
  name: string;
  specialization: string;
  commissionPercent?: number;
  phone?: string;
  avatarInitials?: string;
}

export interface StaffListResponse {
  staff: Staff[];
}

export interface CreateStaffBody {
  name: string;
  specialization: string;
  commissionPercent?: number;
  phone?: string;
}

export interface Service {
  id: string;
  name: string;
  category: string;
  price: number;
  duration: number;
}

export interface ServiceListResponse {
  services: Service[];
  categories: string[];
}

export interface CreateServiceBody {
  name: string;
  category: string;
  price: number;
  duration: number;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  brand?: string;
  stockQuantity: number;
  costPrice?: number;
  sellingPrice: number;
  isLowStock?: boolean;
}

export interface ProductListResponse {
  products: Product[];
}

export interface CreateProductBody {
  name: string;
  category: string;
  brand?: string;
  stockQuantity: number;
  costPrice?: number;
  sellingPrice: number;
  lowStockThreshold?: number;
}

export interface AppointmentListResponse {
  appointments: Appointment[];
}

export interface CreateAppointmentBody {
  customerId?: string;
  staffId: string;
  serviceId: string;
  appointmentDate: string;
  appointmentTime: string;
  notes?: string;
}

export interface UpdateAppointmentBody {
  status?: string;
  notes?: string;
}

export interface BillItem {
  type: string;
  itemId: string;
  name: string;
  staffId?: string;
  staffName?: string;
  price: number;
  quantity: number;
  discount?: number;
  total: number;
}

export interface Bill {
  id: string;
  billNumber: string;
  customerId?: string;
  customerName?: string;
  customerPhone?: string;
  items?: BillItem[];
  subtotal?: number;
  taxPercent?: number;
  taxAmount?: number;
  discountAmount?: number;
  finalAmount: number;
  paymentMethod: string;
  status: string;
  createdAt?: string;
}

export interface BillListResponse {
  bills: Bill[];
}

export interface CreateBillBody {
  customerId?: string;
  customerName?: string;
  customerPhone?: string;
  items: BillItem[];
  subtotal?: number;
  taxPercent?: number;
  taxAmount?: number;
  discountAmount?: number;
  finalAmount: number;
  paymentMethod: string;
  status: string;
}

export interface Expense {
  id: string;
  category: string;
  description?: string;
  amount: number;
  date: string;
  createdAt?: string;
}

export interface ExpenseListResponse {
  expenses: Expense[];
}

export interface CreateExpenseBody {
  category: string;
  description?: string;
  amount: number;
  date: string;
}

export interface Membership {
  id: string;
  name: string;
  price: number;
  duration: number;
  benefits?: string;
  discountPercent?: number;
}

export interface MembershipListResponse {
  memberships: Membership[];
}

export interface CreateMembershipBody {
  name: string;
  price: number;
  duration: number;
  benefits?: string;
  discountPercent?: number;
}

export interface TopServiceItem {
  name: string;
  revenue: number;
  count?: number;
}

export interface TopStaffItem {
  name: string;
  revenue: number;
}

export interface DashboardStats {
  todayRevenue: number;
  todayBills: number;
  todayCustomers?: number;
  totalCustomers: number;
  monthRevenue: number;
  monthBills: number;
  pendingAppointments: number;
  lowStockCount: number;
  topServices?: TopServiceItem[];
  topStaff?: TopStaffItem[];
  recentBills?: Bill[];
}

export interface RevenueDataPoint {
  label: string;
  revenue: number;
  bills?: number;
  customers?: number;
}

export interface RevenueReport {
  period: string;
  totalRevenue: number;
  totalBills: number;
  data: RevenueDataPoint[];
}

export type ListCustomersParams = {
  search?: string;
};

export type ListAppointmentsParams = {
  date?: string;
};

export type GetRevenueReportParams = {
  period?: GetRevenueReportPeriod;
};

export type GetRevenueReportPeriod =
  (typeof GetRevenueReportPeriod)[keyof typeof GetRevenueReportPeriod];

export const GetRevenueReportPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;