# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Salon Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with appointment history
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"

  # ── STAFF ─────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ── BILLS / POS ────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        dob:
          type: string
        notes:
          type: string
        totalSpend:
          type: number
        totalVisits:
          type: number
        createdAt:
          type: string
      required: [id, name, phone]
    CustomerListResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
      required: [customers]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        dob:
          type: string
        notes:
          type: string
      required: [name, phone]
    CustomerDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        dob:
          type: string
        notes:
          type: string
        totalSpend:
          type: number
        totalVisits:
          type: number
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required: [id, name, phone]

    # ── STAFF ──
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        staffId:
          type: string
        staffName:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        serviceCategory:
          type: string
        duration:
          type: number
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required: [id, staffId, serviceId, appointmentDate, appointmentTime, status]
    AppointmentListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required: [appointments]
    CreateAppointmentBody:
      type: object
      properties:
        customerId:
          type: string
        staffId:
          type: string
        serviceId:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: string
      required: [staffId, serviceId, appointmentDate, appointmentTime]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string

    # ── BILL ──
...